requirements.txt
py-clob-client>=0.18.0
requests>=2.28.0
python-dotenv>=1.0.0
